requirements.txt
opencv-python
pytesseract
Pillow
sympy
flask
flask-cors
numpy
requests
google-generativeai